Fixing Filters problem and Refactoring model
Fixed troubles when changin between filters, caused by the problem of not given an unic key to the 'Task' components, refactored the model to work with id instead of the index.

Also added dependencies to the useEffect on App.jsx so the tasks changes on adding, editing or deleting events.